offering a way to force legacy trade by source range
