+++ src/trade-matcher/trade.ts
@@ -1,6 +1,9 @@
 import { ONE, Decimal, BigNumber, BigNumberMax } from '../utils/numerics';
 import { EncodedOrder, DecodedOrder } from '../common/types';
 import { decodeFloat } from '../utils/encoders';
+import { getRuntimeConfig } from '../runtime-config';
+
+const isLegacyTradeBySourceRange = getRuntimeConfig().legacyTradeBySourceRange;
 
 const C = BigNumber.from(ONE);
 
@@ -33,8 +36,12 @@ const getEncodedTradeBySourceAmount = (
   y: BigNumber,
   z: BigNumber,
   A: BigNumber,
-  B: BigNumber
+  B: BigNumber,
+  forceLegacyTradeBySourceRange: boolean = false
 ): BigNumber => {
+  const legacyTradeBySourceRange =
+    forceLegacyTradeBySourceRange || isLegacyTradeBySourceRange;
+
   if (A.eq(0)) {
     return mulDivF(x, mul(B, B), mul(C, C));
   }
@@ -49,7 +56,13 @@ const getEncodedTradeBySourceAmount = (
 
   const temp4 = mulDivC(temp1, temp1, factor);
   const temp5 = mulDivC(temp3, A, factor);
-  if (temp4.add(temp5).lte(MAX_UINT256)) {
+  // newer versions of the CarbonController contract can handle extended range
+  // trades by source. In case the SDK is using an older version of the
+  // contract, handle the trade "same as before" in here.
+  // Before - in a case of overflow it would end up with a 0 rate and not get
+  // picked up for trade. This added check is to avoid the order being picked up
+  // for trade - just to be reverted in the case of overflow by the contract.
+  if (legacyTradeBySourceRange || temp4.add(temp5).lte(MAX_UINT256)) {
     return mulDivF(temp2, temp3.div(factor), temp4.add(temp5));
   }
   return temp2.div(add(A, mulDivC(temp1, temp1, temp3)));
